// src/app/shared/enums/components.name.enum.ts
export enum ComponentsName {
  attendance = 'attendace',
  client = 'client',
  clientType = 'clientType',
  feedback = 'feedback',
  incomeOutcome = 'incomeOutcome',
  material = 'material',
  materialTracking = 'materialTracking',
  note = 'note',
  order = 'order',
  service = 'service',
  servicePricePerClientType = 'servicePricePerClientType',
  serviceType = 'serviceType',
  employee = 'employee',
  shift = 'shift',
  Bank = 'bank',
  supplier = 'supplier',
  commitmentAndDue = 'commitmentAndDue',
  unFinishedOrders = 'unfinishedOrders'
}